+++ relay-event-schema/src/protocol/span/convert.rs
@@ -176,7 +176,8 @@ map_fields!(
     span.data.environment <=> event.environment,
     span.data.browser_name <=> event.contexts.browser.name,
     span.data.sdk_name <=> event.client_sdk.name,
-    span.data.sdk_version <=> event.client_sdk.version
+    span.data.sdk_version <=> event.client_sdk.version,
+    span.origin <=> event.contexts.trace.origin
     ;
     span.is_segment <= true,
     span.was_transaction <= true
@@ -209,6 +210,7 @@ mod tests {
                         "trace_id": "4C79F60C11214EB38604F4AE0781BFB2",
                         "span_id": "FA90FDEAD5F74052",
                         "type": "trace",
+                        "origin": "manual",
                         "op": "myop",
                         "status": "ok",
                         "exclusive_time": 123.4,
@@ -263,7 +265,7 @@ mod tests {
             is_segment: true,
             status: Ok,
             tags: ~,
-            origin: ~,
+            origin: "manual",
             profile_id: EventId(
                 a0aaaaaa-aaaa-aaaa-aaaa-aaaaaaaaaaab,
             ),