+++ relay-server/src/metrics/minimal.rs
@@ -2,9 +2,9 @@ use relay_metrics::{
     BucketMetadata, CounterType, MetricName, MetricNamespace, MetricResourceIdentifier, MetricType,
 };
 use serde::de::IgnoredAny;
-use serde::{de, Deserialize, Deserializer};
+use serde::Deserialize;
 
-use crate::metrics::{BucketSummary, ExtractionMode, TrackableBucket};
+use crate::metrics::{BucketSummary, TrackableBucket};
 
 /// Bucket which parses only the minimally required information to implement [`TrackableBucket`].
 ///
@@ -30,7 +30,7 @@ impl TrackableBucket for MinimalTrackableBucket {
         self.value.ty()
     }
 
-    fn summary(&self, mode: ExtractionMode) -> BucketSummary {
+    fn summary(&self) -> BucketSummary {
         let mri = match MetricResourceIdentifier::parse(self.name()) {
             Ok(mri) => mri,
             Err(_) => return BucketSummary::default(),
@@ -38,10 +38,8 @@ impl TrackableBucket for MinimalTrackableBucket {
 
         match mri.namespace {
             MetricNamespace::Transactions => {
-                let usage = matches!(mode, ExtractionMode::Usage);
                 let count = match self.value {
-                    MinimalValue::Counter(c) if usage && mri.name == "usage" => c.to_f64() as usize,
-                    MinimalValue::Distribution(d) if !usage && mri.name == "duration" => d.0,
+                    MinimalValue::Counter(c) if mri.name == "usage" => c.to_f64() as usize,
                     _ => 0,
                 };
                 let has_profile = matches!(mri.name.as_ref(), "usage" | "duration")
@@ -67,7 +65,7 @@ enum MinimalValue {
     #[serde(rename = "c")]
     Counter(CounterType),
     #[serde(rename = "d")]
-    Distribution(SeqCount),
+    Distribution(IgnoredAny),
     #[serde(rename = "s")]
     Set(IgnoredAny),
     #[serde(rename = "g")]
@@ -91,59 +89,12 @@ struct Tags {
     has_profile: Option<IgnoredAny>,
 }
 
-/// Deserializes only the count of a sequence ingoring all individual items.
-#[derive(Clone, Copy, Debug, Default)]
-struct SeqCount(usize);
-
-impl<'de> Deserialize<'de> for SeqCount {
-    fn deserialize<D>(deserializer: D) -> Result<Self, D::Error>
-    where
-        D: Deserializer<'de>,
-    {
-        struct Visitor;
-
-        impl<'a> de::Visitor<'a> for Visitor {
-            type Value = SeqCount;
-
-            fn expecting(&self, formatter: &mut std::fmt::Formatter) -> std::fmt::Result {
-                formatter.write_str("a sequence")
-            }
-
-            fn visit_seq<A>(self, mut seq: A) -> Result<Self::Value, A::Error>
-            where
-                A: de::SeqAccess<'a>,
-            {
-                let mut count = 0;
-                while seq.next_element::<IgnoredAny>()?.is_some() {
-                    count += 1;
-                }
-
-                Ok(SeqCount(count))
-            }
-        }
-
-        deserializer.deserialize_seq(Visitor)
-    }
-}
-
 #[cfg(test)]
 mod tests {
     use insta::assert_debug_snapshot;
     use relay_metrics::Bucket;
 
     use super::*;
-
-    #[test]
-    fn test_seq_count() {
-        let SeqCount(s) = serde_json::from_str("[1, 2, 3, 4, 5]").unwrap();
-        assert_eq!(s, 5);
-
-        let SeqCount(s) = serde_json::from_str("[1, 2, \"mixed\", 4, 5]").unwrap();
-        assert_eq!(s, 5);
-
-        let SeqCount(s) = serde_json::from_str("[]").unwrap();
-        assert_eq!(s, 0);
-    }
 
     #[test]
     fn test_buckets() {
@@ -214,44 +165,12 @@ mod tests {
 
         for (b, mb) in buckets.iter().zip(min_buckets.iter()) {
             assert_eq!(b.name(), mb.name());
-            assert_eq!(
-                b.summary(ExtractionMode::Usage),
-                mb.summary(ExtractionMode::Usage)
-            );
-            assert_eq!(
-                b.summary(ExtractionMode::Duration),
-                mb.summary(ExtractionMode::Duration)
-            );
+            assert_eq!(b.summary(), mb.summary());
             assert_eq!(b.metadata, mb.metadata);
         }
 
-        let duration = min_buckets
-            .iter()
-            .map(|b| b.summary(ExtractionMode::Duration))
-            .collect::<Vec<_>>();
-        let usage = min_buckets
-            .iter()
-            .map(|b| b.summary(ExtractionMode::Usage))
-            .collect::<Vec<_>>();
-
-        assert_debug_snapshot!(duration, @r###"
-        [
-            Transactions {
-                count: 4,
-                has_profile: true,
-            },
-            Transactions {
-                count: 0,
-                has_profile: false,
-            },
-            Spans(
-                3,
-            ),
-            None,
-            None,
-        ]
-        "###);
-        assert_debug_snapshot!(usage, @r###"
+        let summary = min_buckets.iter().map(|b| b.summary()).collect::<Vec<_>>();
+        assert_debug_snapshot!(summary, @r###"
         [
             Transactions {
                 count: 0,