+++ relay-server/src/services/processor/span/processing.rs
@@ -159,7 +159,7 @@ pub fn process(
 
             let metrics = extract_metrics(
                 span,
-                &CombinedMetricExtractionConfig::new(global_metrics_config, config),
+                CombinedMetricExtractionConfig::new(global_metrics_config, config),
             );
             state.extracted_metrics.project_metrics.extend(metrics);
             item.set_metrics_extracted(true);
@@ -295,13 +295,6 @@ pub fn extract_from_event(
     };
 
     if state.spans_extracted {
-        return;
-    }
-
-    if !state
-        .project_state
-        .has_feature(Feature::ExtractSpansAndSpanMetricsFromEvent)
-    {
         return;
     }
 
@@ -736,7 +729,7 @@ mod tests {
             .config
             .features
             .0
-            .insert(Feature::ExtractSpansAndSpanMetricsFromEvent);
+            .insert(Feature::ExtractCommonSpanMetricsFromEvent);
 
         let event = Event {
             ty: EventType::Transaction.into(),