+++ relay-server/src/services/project/metrics.rs
@@ -101,13 +101,10 @@ impl Buckets<Filtered> {
             })
             .collect();
 
-        let mode = project_state.get_extraction_mode();
-
         if !disabled_namespace_buckets.is_empty() {
             metric_outcomes.track(
                 scoping,
                 &disabled_namespace_buckets,
-                mode,
                 Outcome::Filtered(FilterStatKey::DisabledNamespace),
             );
         }
@@ -116,7 +113,6 @@ impl Buckets<Filtered> {
             metric_outcomes.track(
                 scoping,
                 &denied_buckets,
-                mode,
                 Outcome::Filtered(FilterStatKey::DeniedName),
             );
         }
@@ -130,7 +126,7 @@ fn is_metric_namespace_valid(state: &ProjectState, namespace: MetricNamespace) -
         MetricNamespace::Sessions => true,
         MetricNamespace::Transactions => true,
         MetricNamespace::Spans => {
-            state.has_feature(Feature::ExtractSpansAndSpanMetricsFromEvent)
+            state.has_feature(Feature::ExtractCommonSpanMetricsFromEvent)
                 || state.has_feature(Feature::StandaloneSpanIngestion)
         }
         MetricNamespace::Profiles => true,