+++ tests/integration/consts.py
@@ -1,0 +1,4 @@
+# Minimum supported version for metric transaction by Relay.
+TRANSACTION_EXTRACT_MIN_SUPPORTED_VERSION = 3
+# Maximum supported version for metric transaction by Relay.
+TRANSACTION_EXTRACT_MAX_SUPPORTED_VERSION = 6